+++ src/components/BatchAllocations/BatchAllocationLogDetails/index.js
@@ -1,0 +1,1 @@
+export { BatchAllocationLogDetailsContainer as BatchAllocationLogDetails } from './BatchAllocationLogDetailsContainer';