+++ src/components/BatchAllocations/BatchAllocationsForm/utils.js
@@ -6,16 +6,9 @@ import {
 } from '@folio/stripes-acq-components';
 
 import {
-  BATCH_ALLOCATION_COLUMNS,
   BATCH_ALLOCATION_FIELDS,
   BATCH_ALLOCATION_FORM_SPECIAL_FIELDS,
 } from '../constants';
-
-export const getBatchAllocationColumnMapping = ({ intl }) => {
-  return BATCH_ALLOCATION_COLUMNS.reduce((acc, column) => {
-    return { ...acc, [column]: intl.formatMessage({ id: `ui-finance.transaction.allocation.batch.columns.${column}` }) };
-  }, {});
-};
 
 export const getFormattedOptions = (intl, dataOptions) => dataOptions.map(({ labelId, ...rest }) => ({
   label: intl.formatMessage({ id: labelId }),