+++ src/components/BatchAllocations/utils.js
@@ -1,86 +1,37 @@
-import { FormattedMessage } from 'react-intl';
-
 import {
-  Button,
-  Checkbox,
-  Icon,
-  NoValue,
-  TextLink,
+  exportToCsv,
 } from '@folio/stripes/components';
-import { getFullName } from '@folio/stripes/util';
-import { FolioFormattedTime } from '@folio/stripes-acq-components';
 
 import { BATCH_ALLOCATION_ROUTES_DICT } from '../../common/const';
-import { BATCH_ALLOCATION_LOG_FIELDS } from './BatchAllocationLogs/constants';
-import { BATCH_ALLOCATION_LOG_STATUSES } from './constants';
+import { BATCH_ALLOCATION_COLUMNS } from './constants';
+
+export const getBatchAllocationColumnMapping = ({ intl }) => {
+  return BATCH_ALLOCATION_COLUMNS.reduce((acc, column) => {
+    return { ...acc, [column]: intl.formatMessage({ id: `ui-finance.transaction.allocation.batch.columns.${column}` }) };
+  }, {});
+};
+
+export const exportCsvBatchAllocationLog = (batchAllocationLog, intl) => {
+  const filename = `${batchAllocationLog.jobName}-log`;
+  const exportData = batchAllocationLog.jobDetails?.fyFinanceData?.map(data => ({
+    ...data,
+    transactionTag: data.transactionTag?.tagList?.join(', '),
+  }));
+
+  return exportToCsv(
+    [
+      getBatchAllocationColumnMapping({ intl }),
+      ...(exportData || []),
+    ],
+    {
+      header: false,
+      filename,
+    },
+  );
+};
 
 export const resolveDefaultBackPathname = (sourceType, sourceId) => {
   const pathname = `${BATCH_ALLOCATION_ROUTES_DICT[sourceType]}/${sourceId}/view`;
 
   return pathname;
-};
-
-export const getLogsListColumnMapping = ({
-  intl,
-  isAllSelected,
-  selectAll,
-  disabled,
-}) => {
-  const mapping = Object.keys(BATCH_ALLOCATION_LOG_FIELDS).reduce((acc, column) => {
-    return { ...acc, [column]: intl.formatMessage({ id: `ui-finance.allocation.batch.logs.columns.${column}` }) };
-  }, {});
-
-  return {
-    select: (
-      <Checkbox
-        aria-label={intl.formatMessage({ id: 'ui-finance.allocation.batch.logs.columns.selectAll' })}
-        checked={isAllSelected}
-        onChange={selectAll}
-        type="checkbox"
-        disabled={disabled}
-      />
-    ),
-    ...mapping,
-  };
-};
-
-export const getResultsListFormatter = ({
-  disabled,
-  intl,
-  selectRecord,
-  selectedRecordsMap,
-}) => {
-  return {
-    select: (item) => (
-      <Checkbox
-        aria-label={intl.formatMessage({ id: 'ui-finance.allocation.batch.logs.columns.select' }, { term: item.id })}
-        type="checkbox"
-        checked={Boolean(selectedRecordsMap[item.id])}
-        onChange={() => selectRecord(item)}
-        disabled={disabled}
-      />
-    ),
-    [BATCH_ALLOCATION_LOG_FIELDS.jobName]: (item) => (
-      <>
-        <TextLink to="#"> {item.jobName} </TextLink>
-        <Button
-          onClick={() => {}}
-          buttonStyle="none"
-          bottomMargin0
-        >
-          <Icon size="small" icon="download" />
-        </Button>
-      </>
-    ),
-    [BATCH_ALLOCATION_LOG_FIELDS.status]: (item) => (
-      <FormattedMessage id={`ui-finance.allocation.batch.logs.columns.status.${BATCH_ALLOCATION_LOG_STATUSES[item.status]}`} />
-    ),
-    [BATCH_ALLOCATION_LOG_FIELDS.recordsCount]: (item) => item.recordsCount || <NoValue />,
-    [BATCH_ALLOCATION_LOG_FIELDS.createdDate]: (item) => <FolioFormattedTime dateString={item.metadata.createdDate} />,
-    [BATCH_ALLOCATION_LOG_FIELDS.updatedDate]: (item) => <FolioFormattedTime dateString={item.metadata.updatedDate} />,
-    createdByUsername: (item) => (
-      <TextLink to={`/users/preview/${item.createdByUser.id}`}> {getFullName(item.createdByUser)}</TextLink>
-    ),
-    [BATCH_ALLOCATION_LOG_FIELDS.jobNumber]: (item) => (item.jobNumber || <NoValue />),
-  };
 };